Backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
channels==4.1.0
django-cors-headers==4.3.1
requests==2.31.0
Pillow==11.2.1
python-decouple==3.8
python-dotenv==1.0.1
